--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,27 +17,34 @@ edition = "2021"
 # Command-line
 clap = { version = "3", default-features = false, features = ["std", "cargo"] }
 # Server
-tokio = { version = "1", features = ["rt", "macros"]}
-hyper = { version = "0.14.10", features = ["http1", "server", "tcp"] }
+tokio = { version = "1", features = ["fs", "macros", "io-util", "rt"] }
+tokio-util = { version = "0.7", features = ["io"] }
+hyper = { version = "0.14", features = ["http1", "server", "stream", "tcp"] }
 headers = "0.3"
 mime_guess = "2.0"
 percent-encoding = "2.1"
-# Compression
-brotli = "3"
-flate2 = "1.0"
 # Rendering
 tera = "1"
-serde = { version = "1.0", features = ["derive"] } # For tera serializing variables to template.
+serde = { version = "1.0", features = [
+  "derive",
+] } # For tera serializing variables to template.
 ignore = "0.4" # Respect to .gitignore while listing directories.
 # Logging
 chrono = "0.4"
 # Directory Download
 qstring = "0.7"
 zip = { version = "0.5", default-features = false, features = ["deflate"] }
-
-[dev-dependencies]
 tempfile = "3"
+sysinfo = "0.25"
 once_cell = "1"
+futures = "0.3"
+bytes = "1"
+async-compression = { version = "0.3", features = [
+  "brotli",
+  "deflate",
+  "gzip",
+  "tokio",
+] }
 
 [profile.dev]
 codegen-units = 2