closes #33 

# TODO

- [x] stream compression
  - https://lib.rs/crates/async-compression
- [ ] preserve performance