@henry40408 2022-09-01
This pull request introduces a huge performance loss:

`master` branch 7113041

```
Running 10s test @ http://localhost:5000/Cargo.toml
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.96ms  460.61us  16.29ms   94.71%
    Req/Sec     1.70k    67.43     1.93k    84.00%
  33916 requests in 10.01s, 51.78MB read
Requests/sec:   3388.94
Transfer/sec:      5.17MB
```

pull request #94 a8298b9

```
Running 10s test @ http://localhost:5000/Cargo.toml
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    44.61ms    6.18ms  63.94ms   94.81%
    Req/Sec   112.14     11.46   166.00     71.00%
  2236 requests in 10.01s, 3.86MB read
Requests/sec:    223.35
Transfer/sec:    395.05KB
```